use 32bit ints for benchmark make hash function variadic use xxhash as hash function add hash instruction test

+++ benchmark/include/DateReformatter.hpp
@@ -12,5 +12,5 @@ class DateReformatter
   public:
     explicit DateReformatter(std::vector<std::string> &&column);
 
-    std::vector<int64_t> reformat();
+    std::vector<int32_t> reformat();
 };
